requirements.txt
contractions==0.1.73
optimum==1.20.0
pandas==2.2.2
praw==7.7.1
PyYAML==6.0.1
transformers==4.41.2
streamlit==1.35.0